test: move git mock
