+++ lib/modules/manager/helmfile/artifacts.spec.ts
@@ -2,7 +2,7 @@ import { codeBlock } from 'common-tags';
 import { mockDeep } from 'vitest-mock-extended';
 import { join } from 'upath';
 import { envMock, mockExecAll } from '../../../../test/exec-util';
-import { env, fs, git, mocked } from '../../../../test/util';
+import { env, fs, mocked } from '../../../../test/util';
 import { GlobalConfig } from '../../../config/global';
 import type { RepoGlobalConfig } from '../../../config/types';
 import * as docker from '../../../util/exec/docker';
@@ -10,6 +10,7 @@ import * as hostRules from '../../../util/host-rules';
 import * as _datasource from '../../datasource';
 import type { UpdateArtifactsConfig } from '../types';
 import * as helmfile from '.';
+import { git } from '../../../../test/git';
 
 vi.mock('../../datasource', () => mockDeep());
 vi.mock('../../../util/exec/env');
