+++ lib/modules/manager/nuget/artifacts.spec.ts
@@ -1,7 +1,7 @@
 import { mockDeep } from 'vitest-mock-extended';
 import { join } from 'upath';
 import { envMock, mockExecAll } from '../../../../test/exec-util';
-import { env, fs, git, mocked, scm } from '../../../../test/util';
+import { env, fs, mocked, scm } from '../../../../test/util';
 import { GlobalConfig } from '../../../config/global';
 import type { RepoGlobalConfig } from '../../../config/types';
 import * as docker from '../../../util/exec/docker';
@@ -10,6 +10,7 @@ import type { UpdateArtifactsConfig } from '../types';
 import type { Registry } from './types';
 import * as util from './util';
 import * as nuget from '.';
+import { git } from '../../../../test/git';
 
 vi.mock('../../../util/exec/env');
 vi.mock('../../../util/fs');
