+++ lib/util/git/set-branch-commit.spec.ts
@@ -1,4 +1,5 @@
-import { git, logger, mocked, partial } from '../../../test/util';
+import { git } from '../../../test/git';
+import { logger, mocked, partial } from '../../../test/util';
 import * as _repositoryCache from '../cache/repository';
 import type { BranchCache, RepoCacheData } from '../cache/repository/types';
 import { setBranchNewCommit } from './set-branch-commit';
