+++ lib/workers/repository/config-migration/branch/rebase.spec.ts
@@ -1,12 +1,13 @@
 import type { Indent } from 'detect-indent';
 import JSON5 from 'json5';
 import { Fixtures } from '../../../../../test/fixtures';
-import { RenovateConfig, git, partial, scm } from '../../../../../test/util';
+import { RenovateConfig, partial, scm } from '../../../../../test/util';
 import { getConfig } from '../../../../config/defaults';
 import { GlobalConfig } from '../../../../config/global';
 import { MigratedDataFactory } from './migrated-data';
 import type { MigratedData } from './migrated-data';
 import { jsonStripWhitespaces, rebaseMigrationBranch } from './rebase';
+import { git } from '../../../../../test/git';
 
 vi.mock('../../../../util/git');
