+++ lib/workers/repository/onboarding/branch/check.spec.ts
@@ -1,6 +1,6 @@
+import { git } from '../../../../../test/git';
 import {
   RenovateConfig,
-  git,
   mocked,
   partial,
   platform,
