+++ lib/workers/repository/onboarding/pr/pr-list.spec.ts
@@ -51,7 +51,7 @@ describe('workers/repository/onboarding/pr/pr-list', () => {
         <details>
         <summary>Lock file maintenance</summary>
 
-          - Schedule: ["before 5am"]
+          - Schedule: [\\"before 5am\\"]
           - Branch name: \`renovate/lock-file-maintenance\`
           - Merge into: \`base\`
           - Regenerate lock files to use latest dependency versions
