+++ lib/workers/repository/process/index.spec.ts
@@ -1,6 +1,5 @@
 import {
   RenovateConfig,
-  git,
   logger,
   mocked,
   platform,
@@ -14,6 +13,7 @@ import { getCache } from '../../../util/cache/repository';
 import * as _extractUpdate from './extract-update';
 import { lookup } from './extract-update';
 import { extractDependencies, updateRepo } from '.';
+import { git } from '../../../../test/git';
 
 vi.mock('../../../util/git');
 vi.mock('./extract-update');
