+++ lib/workers/repository/update/branch/get-updated.spec.ts
@@ -1,4 +1,5 @@
-import { git, mocked } from '../../../../../test/util';
+import { git } from '../../../../../test/git';
+import { mocked } from '../../../../../test/util';
 import { GitRefsDatasource } from '../../../../modules/datasource/git-refs';
 import * as _batectWrapper from '../../../../modules/manager/batect-wrapper';
 import * as _bundler from '../../../../modules/manager/bundler';
