+++ lib/workers/repository/update/branch/index.spec.ts
@@ -1,11 +1,4 @@
-import {
-  fs,
-  git,
-  mocked,
-  partial,
-  platform,
-  scm,
-} from '../../../../../test/util';
+import { fs, mocked, partial, platform, scm } from '../../../../../test/util';
 import { getConfig } from '../../../../config/defaults';
 import { GlobalConfig } from '../../../../config/global';
 import type { RepoGlobalConfig } from '../../../../config/types';
@@ -44,6 +37,7 @@ import * as _getUpdated from './get-updated';
 import * as _reuse from './reuse';
 import * as _schedule from './schedule';
 import * as branchWorker from '.';
+import { git } from '../../../../../test/git';
 
 vi.mock('./get-updated');
 vi.mock('./schedule');
