+++ lib/workers/repository/update/pr/code-owners.spec.ts
@@ -1,8 +1,9 @@
 import { codeBlock } from 'common-tags';
 import { mock } from 'vitest-mock-extended';
-import { fs, git } from '../../../../../test/util';
+import { fs } from '../../../../../test/util';
 import type { Pr } from '../../../../modules/platform';
 import { codeOwnersForPr } from './code-owners';
+import { git } from '../../../../../test/git';
 
 vi.mock('../../../../util/fs');
 vi.mock('../../../../util/git');
